Week 4/presidio_internal_research_agent/requirements.txt
langchain>=0.1.16
langchain-community
langchain-aws
boto3
tavily-python
python-dotenv
langchain-tavily
chromadb
langchain-text-splitters
sentence-transformers

fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
